# sstable: implement NextPrefix for the sstable iterators