--- a/sstable/raw_block.go
+++ b/sstable/raw_block.go
@@ -26,7 +26,8 @@ func (w *rawBlockWriter) add(key InternalKey, value []byte) {
 	w.curKey = w.curKey[:size]
 	copy(w.curKey, key.UserKey)
 
-	w.storeWithOptionalValuePrefix(size, value, false, 0, false)
+	w.storeWithOptionalValuePrefix(
+		size, value, len(key.UserKey), false, 0, false)
 }
 
 // rawBlockIter is an iterator over a single block of data. Unlike blockIter,