--- a/sstable/value_block.go
+++ b/sstable/value_block.go
@@ -132,12 +132,13 @@ const (
 	valueKindIsInPlaceValue valuePrefix = '\x00'
 
 	// 1 bit indicates SET has same key prefix as immediately preceding key that
-	// is also a SET. This is optional, in that if this bit is false, there is
-	// no information.
+	// is also a SET. If the immediately preceding key in the same block is a
+	// SET, AND this bit is 0, the prefix must have changed.
 	//
 	// Note that the current policy of only storing older MVCC versions in value
 	// blocks means that valueKindIsValueHandle => SET has same prefix. But no
-	// code should rely on this behavior.
+	// code should rely on this behavior. Also, SET has same prefix does *not*
+	// imply valueKindIsValueHandle.
 	setHasSameKeyPrefixMask valuePrefix = '\x20'
 
 	// 3 least-significant bits for the user-defined base.ShortAttribute.