The NextPrefix implementation mostly delegates to blockIter.nextPrefix.
If the next-prefix is not found in the current data block they first
attempt to step the index block and see if the successor key is
contained in that block. If that fails, they fall back to an approach
similar to SeekGE by also seeking the index.

blockIter.nextPrefix is defined for both v2 and v3 sstables, with the
latter utilizing the setHasSameKeyPrefix bit encoded in the value
prefix, and the bit encoded in the key restarts. To utilize this bit
efficiently, the shared prefix (for key prefix compression) when
writing a v3 sstable is limited to the prefix length of the preceding
key. This allows the read path to step over many keys that share
the same prefix, without assembling the full key, since any of them
can subsequently service the key prefix decompression.

The IteratorScanNextPrefix benchmark uses both NextPrefix and SeekGE,
to evaluate the difference. In general, NextPrefix is faster.
Additionally, the v3 format is faster than the v2 format when doing
NextPrefix. See the code comment in the benchmark with a more detailed
analysis.

The old is v2 and new is v3.

```
name                                                       old time/op    new time/op    delta
IteratorScanNextPrefix/versions=1/method=seek-ge-10          52.8ns ± 0%    54.4ns ± 0%   +3.04%  (p=0.008 n=5+5)
IteratorScanNextPrefix/versions=1/method=next-prefix-10      33.3ns ± 0%    15.4ns ± 0%  -53.62%  (p=0.016 n=4+5)
IteratorScanNextPrefix/versions=2/method=seek-ge-10          82.5ns ± 0%    90.0ns ± 0%   +9.21%  (p=0.008 n=5+5)
IteratorScanNextPrefix/versions=2/method=next-prefix-10      61.6ns ± 0%    19.1ns ± 0%  -68.95%  (p=0.008 n=5+5)
IteratorScanNextPrefix/versions=10/method=seek-ge-10          828ns ± 0%     936ns ± 2%  +13.13%  (p=0.008 n=5+5)
IteratorScanNextPrefix/versions=10/method=next-prefix-10      495ns ± 1%      63ns ± 0%  -87.33%  (p=0.008 n=5+5)
IteratorScanNextPrefix/versions=100/method=seek-ge-10        1.22µs ± 0%    1.04µs ± 0%  -15.02%  (p=0.016 n=4+5)
IteratorScanNextPrefix/versions=100/method=next-prefix-10    1.03µs ± 1%    0.27µs ± 1%  -73.95%  (p=0.008 n=5+5)

name                                                       old alloc/op   new alloc/op   delta
IteratorScanNextPrefix/versions=1/method=seek-ge-10           0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=1/method=next-prefix-10       0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=2/method=seek-ge-10           0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=2/method=next-prefix-10       0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=10/method=seek-ge-10          0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=10/method=next-prefix-10      0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=100/method=seek-ge-10         0.00B          0.00B          ~     (all equal)
IteratorScanNextPrefix/versions=100/method=next-prefix-10     0.00B          0.00B          ~     (all equal)
```

Informs https://github.com/cockroachdb/pebble/issues/1170